[runtime] Reduce the race window when aborting Monitor:Enter.
If you abort a thread that is doing a Monitor:Enter, if the syscall abort lands before
the target thread start waiting it will block in a non aletable way.

To address that, we start by waiting in an alertable way and if interrupted we
decide to continue or not based on whether an abort can start.

Additionally, when alertable, we CAS the thread state to catch early Thread:Abort
calls that set the state by did not managed to init it.

This fixes a rare deadlock on tests/monitor-abort.exe.


Commit migrated from https://github.com/mono/mono/commit/60e0a82be212fb506551812af44cf0865f89e7f0