[os-event] Fix set operation
This is a logic bug where we would only wake up one of the threads that are waiting for the manual MonoOSEvent to be set. This would be a problem only if more than one thread is waiting.


Commit migrated from https://github.com/mono/mono/commit/627eaebf94e6218891656fbda6ed722226cc2f59