[runtime] Add new thread state INTERRUPT_REQUEST_DEFERRED to fix race with abort protected blocks.
The way this works is that when we install a protected block, we change a interrupt requested to interrupt request deferred.
Then, when protected block count drops to zero, we switch back and inform the caller to poll for interrupts.

Additionally, when interrupting, we set it to deferred if there's a protected block in effect and it's not a self interrupt.


Commit migrated from https://github.com/mono/mono/commit/5c7d98989a0746238c0d5c3f7d0c16bdff77b084