[sgen] Don't assert when suspending detaching thread (mono/mono#4292)
This means that the thread will be part of the stw (not skipped), so if the stack pointer is not between the expected ranges we dont scan the thread. In other parts of the runtime, suspending here is handled by the suspend_can_continue flag.

Commit migrated from https://github.com/mono/mono/commit/17c867e883000ec7f5b22c3449f07f962a0273cf