[threadpool] Fix hang on threadpool cleanup (mono/mono#4330)
Waiting for the threadpool threads to exit on shutdown triggers a hand. The simplest solution is to not wait for them, and let the OS clean them up.

This is not an ideal solution, but this is to fix a months-long chase for this bug, and to release a less-buggy monthly release.

Commit migrated from https://github.com/mono/mono/commit/ede498bb6fc08d45c8eaeafb2be08ea3c8e1066a