[threads] Don't ignore abort requests in abort protected blocks
Either interrupt or set the interrupt_deferred flag, so the thread can handle the request later in the future. Otherwise we are left with a thread that has an AbortRequested set on its state that it is never processed and future abort requests are ignored due to the presence of this flag.


Commit migrated from https://github.com/mono/mono/commit/bc25f308cabad9bb333a1d596fe89881205cc82f